DS-1815 Enable multi-classification in Logistic Regression

+++ deeplang/src/main/scala/io/deepsense/deeplang/doperables/machinelearning/logisticregression/LogisticRegressionParameters.scala
@@ -17,4 +17,4 @@
 package io.deepsense.deeplang.doperables.machinelearning.logisticregression
 
 case class LogisticRegressionParameters(
-    regularization: Double, iterationsNumber: Double, tolerance: Double)
+    regularization: Double, iterationsNumber: Int, tolerance: Double)
